component structure and placement changed
