+++ web/satellite/src/components/dashboard/newProject/NewProjectArea.vue
